🗺️ Atlas: [extract assembler implementation]
