--- find_long_fns.py
@@ -1,12 +1,0 @@
-import ast
-import re
-import sys
-
-def find_long_functions(filepath):
-    with open(filepath, 'r') as f:
-        content = f.read()
-
-    # Simple regex to find fn ... {
-    # Not perfect but gives an idea
-    # Let's use a rust parser or a simpler script to find long functions
-    pass
